メインプログラムをクラス化した
