+++ 07.ls_object/main.rb
@@ -1,0 +1,5 @@
+# run_ls.rb
+require_relative 'ls_command'
+
+ls_command = LsCommand.new
+ls_command.run
